+++ cli/src/cli.rs
@@ -2146,7 +2146,6 @@ mod tests {
                 "lamports": 50,
                 "owner": "11111111111111111111111111111111",
                 "executable": false,
-                "rentEpoch": 1,
             }),
         });
         let mut mocks = HashMap::new();
@@ -2186,7 +2185,6 @@ mod tests {
                 "lamports": 42,
                 "owner": "Vote111111111111111111111111111111111111111",
                 "executable": false,
-                "rentEpoch": 1,
             }),
         });
         let mut mocks = HashMap::new();
@@ -2249,7 +2247,6 @@ mod tests {
                 "lamports": 50,
                 "owner": "11111111111111111111111111111111",
                 "executable": false,
-                "rentEpoch": 1,
             }),
         });
         let mut mocks = HashMap::new();