+++ CHANGELOG.md
@@ -29,6 +29,7 @@ Release channels have their own copy of this changelog:
 #### Breaking
 * Added a `slot` property to `EpochRewardsPeriodActiveErrorData`
 * Added error data containing a `slot` property to `RpcCustomError::SlotNotEpochBoundary`
+* Removed the `rent_epoch` property from `UiAccount`
 
 #### Changes
 * The subscription server now prioritizes processing received messages before sending out responses. This ensures that new subscription requests and time-sensitive messages like `PING` opcodes take priority over notifications.